+++ packages/rust/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Now the `typing` module has `std` feature which can be enabled to have automatic context clear logic.
 - Now 9x faster over the JS version (previosuly 7x, ~63ms -> ~51ms)
 - The wasm build is also smaller (as no `std` bloat)
+- **Major Change** :- Now Custom Options are accepted via a struct (CustomOptions) instead of of a HashMap.
 
 ## rust-lib@1.1.0